Update User Reducer to hold all user information from token payload globally
This commits stores all user information from the token payload inside the
user reducer. As the react-oidc-contect only stores the default claims the
token gets additonally decoded and if existing the admin role gets stored
in the reducer. The admin logic is not yet in use as the default role is
admin, but will be in the future.