Switch to managed-nfs-storage storage class to prevent mounting errors
This commit switches the storage class for all tekton pipeline runs for
the frontend repository. The default storage class in the current cluster
uses the nfs csi driver that seems to have problems mounting the pvc workspace
simultaniously to all the tekton pods. This resulted in a 20 minutes pipeline
run instead of 5 min. To prevent this managed-nfs-storage is used again.